Moved bound implementations into the corresponding modules
